% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AEQcalc.R
\name{AEQcalc}
\alias{AEQcalc}
\title{SUB AEQcalc}
\usage{
AEQcalc(input)
}
\value{
AEQ (scalar)
}
\description{
Compute the AEQs for each age group.  Original VB code:
}
\details{
TAB: what are AEQs? Adult Equivalents!
Sub AEQcalc()
Dim TmpA As Double  'TAB: added line
Dim TmpS As Double  'TAB: added line
Dim Age As Integer   'TAB: added line
TmpA = 0
TmpS = 0
For Age = MaxAge% To MinAge% Step -1
AEQ(Age) = MatRate(Age) + TmpS * (1 - MatRate(Age)) * TmpA
TmpA = AEQ(Age)
TmpS = 1 - NatMort(Age)
Next Age

End Sub
}
