% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autocorrel.R
\name{Autocorrel}
\alias{Autocorrel}
\title{Autocorrel}
\usage{
Autocorrel(p, lastx, x)
}
\arguments{
\item{p}{autocorrelation}

\item{lastx}{last value of variable x}

\item{x}{A random (non-correlated) variable generated from the gamma function for the variable.}
}
\value{
New random autocorrelated variable (scalar)
}
\description{
Compute autocorrelated variable, p = autocorrelation, lastx = last value of variable x
}
