% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BufferInit.R
\name{BufferInit}
\alias{BufferInit}
\title{BufferInit}
\usage{
BufferInit(Buffer, inputs)
}
\arguments{
\item{Buffer}{Scaling factor for Target ER}

\item{inputs}{Inputs read from the .rav file}
}
\value{
list with new target rate for the simulation if StepFunc=ER or new capacity (b parameter in SR function) if StepFunc=Pop.
}
\description{
ADJUST TARGET RATE FOR BUFFER EXPLOITATION
}
\details{
If the StepFunc is ER, then the simulation steps through different exploitation rates.
It does this by using the Target ER (in inputs$TargetU and multiplying that by a 'buffer' or scaling factor)
}
