% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompAgeCohort.R
\name{CompAgeCohort}
\alias{CompAgeCohort}
\title{CompAgeCohort}
\usage{
CompAgeCohort(TempCohort, Cohort, inputs)
}
\value{
Cohort (scalar)
EEH: this only works if MinAge=2, because Cohort[1] was set in CompRecruits (not Cohort[minage-1])
EEH: why isn't TempCohort[inputs$MinAge - 1] = Cohort[inputs$MinAge - 1]?
}
\description{
Compute age cohort
}
\details{
Original VB code:
Sub CompAgeCohort(TempCohort() As Double)
    Dim Age As Integer      'counter for ages

For Age = MaxAge% To MinAge% + 1 Step -1
Cohort(Age) = TempCohort(Age - 1)
Next Age

Cohort(MinAge%) = Cohort(MinAge% - 1)

End Sub
}
