% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompEscpmnt.R
\name{CompEscpmnt}
\alias{CompEscpmnt}
\title{CompEscpmnt}
\usage{
CompEscpmnt(Regime, Year, inputs, BufTargetU, Cohort, AEQ, YearStats)
}
\arguments{
\item{Regime}{Harvest regime.}

\item{Year}{Year to compute the escapement for.}

\item{inputs}{Inputs from .rav file}

\item{BufTargetU}{Target ER to use for simulation.  Adjusted if StepFunc=ER.}

\item{Cohort}{Cohort}

\item{AEQ}{Adult equivalents}

\item{YearStats}{list of computed variables for each year: AEQMort, Escpmnt[Year,] = Escpmnt, TotAdultEscpmnt, TotAEQMort, TotEscpmnt,TempCohort.}
}
\value{
Updated YearStats list for value of variables in Year
}
\description{
Compute escapement
}
