% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompNatMort.R
\name{CompNatMort}
\alias{CompNatMort}
\title{CompNatMort}
\usage{
CompNatMort(inputs, CohortBeforeNatMort)
}
\arguments{
\item{inputs}{Inputs from .rav file}

\item{CohortBeforeNatMort}{Cohort before natural mortality}
}
\value{
Cohort after natural mortality
}
\description{
Let the number of fish in each age class decrease according to
the natural mortality in that age class.
}
\details{
Original VB code:
Sub CompNatMort()
    Dim Age%     'counter for ages

For Age% = MinAge% - 1 To MaxAge%
Cohort(Age%) = Cohort(Age%) * (1 - NatMort(Age%))
Next Age%

End Sub
}
