% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompRecruits.R
\name{CompRecruits}
\alias{CompRecruits}
\title{CompRecruits}
\usage{
CompRecruits(YearStats, Year, inputs, repvars, staticvars, BufSRb)
}
\arguments{
\item{YearStats}{list of computed variables for each year: AEQMort, Escpmnt[Year,] = Escpmnt, TotAdultEscpmnt, TotAEQMort, TotEscpmnt,TempCohort.}

\item{Year}{Year to compute the escapement for.}

\item{inputs}{Inputs from .rav file}

\item{repvars}{repvars}

\item{staticvars}{Static variables}

\item{BufSRb}{Capacity.  Changes if StepFunc=Pop.  Otherwise stays the same.}
}
\value{
updated repvars list with: Cohort[1]=CohortAge1, LastRanFlow, LastRanError, LastRanMarine.
}
\description{
Compute recruits
}
