% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompStatsEEH.R
\name{CompStatsEEH}
\alias{CompStatsEEH}
\title{CompStats}
\usage{
CompStatsEEH(BufNum, inputs, BufSRb, YearStats, SummaryStats)
}
\arguments{
\item{BufNum}{Which simulation (TargetER or Pop) is this for.}

\item{inputs}{Inputs from .rav file}

\item{BufSRb}{Capacity.  Changes if StepFunc=Pop.  Otherwise stays the same.}

\item{YearStats}{list of computed variables for each year: AEQMort, Escpmnt[Year,] = Escpmnt, TotAdultEscpmnt, TotAEQMort, TotEscpmnt,TempCohort.}

\item{SummaryStats}{list with the summary statistics to be updated}
}
\value{
Updated SummaryStats list
}
\description{
Returns the statistics (calculated values) needed to produce the summary output file
}
\details{
This function similar but not identical to the original VB function
}
