% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompStockStatus.R
\name{CompStockStatus}
\alias{CompStockStatus}
\title{CompStockStatus}
\usage{
CompStockStatus(TotAdultEscpmnt, inputs)
}
\arguments{
\item{TotAdultEscpmnt}{Total adult escapement}

\item{inputs}{Inputs from .rav file}
}
\value{
Status ("True"/"False")
}
\description{
Compute stock status: is TotAdultEscpmnt > EscpmntBreakPoint[Break]
}
\details{
Original VB code
 Sub CompStockStatus(TotAdultEscpmnt As Double, Status%)
    Dim Break%
    Dim EscpmntCheck$
    
    Break% = 1
    EscpmntCheck$ = "True"
    While Break% <= NumBreakPoints% And EscpmntCheck$ = "True"
        If TotAdultEscpmnt > EscpmntBreakPoint(Break%) Then
            EscpmntCheck$ = "True"
            Break% = Break% + 1
        Else
            EscpmntCheck$ = "False"
        End If
    Wend

    Status% = Break%

End Sub
}
