% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeRER.R
\name{ComputeRER}
\alias{ComputeRER}
\title{ComputeRER}
\usage{
ComputeRER(VRAPList, UEL = 80, LEL = 5)
}
\arguments{
\item{VRAPList}{output list from Main() or RunSims()}

\item{UEL}{Upper target as percent of VRAP simulations that should be above recovery threshold}

\item{LEL}{lower target as percent of VRAP simulations that should be not hit the lower threshold}
}
\value{
a list with:
\describe{
 \item{ERUEL}{The ER at the UEL}
 \item{ERLEL}{The ER at the LEL}
 \item{hr.vrap}{The harvest rates for each simulation in the VRAP output.}
 \item{hr.smooth}{The harvest rates used to computed the smoothed UEL and LEL.}
 \item{uel.vrap}{The UEL for each simulation in the VRAP output.}
 \item{uel.smooth}{The smoothed UEL.}
 \item{lel.vrap}{The LEL for each simulation in the VRAP output.}
 \item{lel.smooth}{The smoothed LEL.}
 \item{UEL}{The proportion to use for the upper threshold.}
 \item{LEL}{The proportion to use for the lower threshold.}
}
}
\description{
Computes the LEL and UEL RERs=
}
\details{
This fits a spline to the harvest versus UET and LET lines to get the ER at the LEL and UEL.
}
