% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaSample.R
\name{GammaSample}
\alias{GammaSample}
\title{GammmaSample}
\usage{
GammaSample(Alpha, Beta)
}
\arguments{
\item{Alpha}{alpha parameter of gamma}

\item{Beta}{beta parameter of gamma}
}
\value{
Gamma distributed random variable (scalar)
}
\description{
Function generates a random gamma deviate with shape
paramater alpha and scale parameter beta
}
