% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Main}
\alias{Main}
\title{Main}
\usage{
Main(InFile = NULL, OutFileBase = NULL, NRuns = -1, NYears = -1,
  Title = -1, TargetStart = -1, TargetEnd = -1, TargetStep = -1,
  ERecovery = -1, QET = -1, ECrit = -1, NewRavFileName = "tmprav.rav",
  forceNewRav = NULL, silent = FALSE, lcores = 1)
}
\arguments{
\item{InFile}{The name of the .rav file}

\item{OutFileBase}{The basename for the .sum, .byr, and .esc output files}

\item{NRuns}{Number of runs to use in the simulations if the user wants to use something different than what is in the .rav file}

\item{NYears}{Number of years to project forward in the simulations if the user wants to use something different than what is in the .rav file}

\item{Title}{Title to use for the report if the user wants to use something different than what is in the .rav file}

\item{TargetStart}{Target ER to start simulations at if the user wants to use something different than what is in the .rav file}

\item{TargetEnd}{Target ER to end simulations at if the user wants to use something different than what is in the .rav file}

\item{TargetStep}{Target ER step sizes if the user wants to use something different than what is in the .rav file}

\item{ERecovery}{Recovery target if the user wants to use something different than what is in the .rav file}

\item{QET}{if the user wants to use something different than what is in the .rav file}

\item{ECrit}{if the user wants to use something different than what is in the .rav file}

\item{NewRavFileName}{A new .rav file is saved in case the user has changed any values from what is in the .rav file.}

\item{forceNewRav}{Force use fo new rav file.  Needed for shiny app.}

\item{silent}{Whether to show progress bar.}

\item{lcores}{Number of cores to use.  Default is non-parallel so lcores=1}
}
\value{
list with output list from RunSims() and output time
}
\description{
Runs VRAP. This function is largely specific to the R version of VRAP.
}
