% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruits.R
\name{Recruits}
\alias{Recruits}
\title{Recruits}
\usage{
Recruits(inputs)
}
\arguments{
\item{inputs}{Inputs from .rav file}
}
\value{
Recruits at age 1
}
\details{
Function Recruits() As Double
Compute factor to convert calculated spawner equivalent
production to age cohort (source is PSC Chinook Model).

Tmp = 0
X9 = 1 - NatMort(1)
For Age% = MinAge% To MaxAge%
X9 = X9 * (1 - NatMort(Age%))
Tmp = Tmp + X9 * MatRate(Age%)
X9 = X9 * (1 - MatRate(Age%))
Next Age%
Recruits = Tmp

End Function
The SR function gets us the AEQRecruits from spawners in year t.
We needs to then translate that to age 1 indiv in pop (Cohort[1])
We know AEQRecruit.  How many Age 1 individuals does that translate to?
Age1 * (1- total fraction lost) = AEQRecruits
So Age1 = AEQRecruits/(1-total fraction lost)
Tmp here is total fraction of age 1 ind that eventually return
AEQRecruits/Tmp = Age 1 or Cohort[1]
}
