% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepInit.R
\name{RepInit}
\alias{RepInit}
\title{RepInit}
\usage{
RepInit(inputs)
}
\arguments{
\item{inputs}{Inputs from .rav file
SET INITIAL SEED FOR AUTOCORRELATED RESIDUALS
SET INITIAL VALUES FOR LAST RANMARINE AND RANFLOW
EEH: VB code was misisng HOC3 and BEV3 from this}
}
