% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSims.R
\name{RunSims}
\alias{RunSims}
\title{RunSims}
\usage{
RunSims(inputs, silent)
}
\arguments{
\item{inputs}{Inputs from .rav file}

\item{silent}{Whether to show progress bar}
}
\value{
list with inputs, SummaryStats, staticvars, comp.time.
CALCULATE AEQs
staticvars are computed variables; they are static
COMPUTE FACTOR TO TRANSLATE AEQ RECRUITMENT TO AGE 1
INITIALIZE REPETITION SPECIFIC PARAMETERS AND ARRAYS
repvar is a list of variables that change each year:
Cohort, LastRanError, LastRanFlow, LastRanMarine
These are updated with each year
starts Cohort at the init # at each age in the input file
Cohort seems to mean population #s at each age
YearStats is a list of variables that are saved for each year
SummaryStats passed in to fix first year Ran to be same across reps
BEGIN Year LOOP
NATURAL MORTALITY PROCESS
Update Cohort (pop size) at Age with Natural Mortality
COMPUTE ESCAPEMENT USING BASE LEVEL EXPLOITATION RATE
CHECK STOCK STATUS
Going to be a number between 1 and NumBreaks
ADJUST REGIME IF WARRANTED
AGE COHORT
now update cohort (non-spawner pop size at age) for next year
using the updated "cohort" from CompEscpmnt
COMPUTE RECRUITS FROM ESCAPEMENT
Uses SR (Escpmnt to AEQRecruits) function with error added to that to get AEQRecruits from this years spawners (Escpmnt)
Needs to then translate that to age 1 indiv in pop (Cohort[1])
We know AEQRecruit.  How many Age 1 individuals does that translate to?  Age1 * (1- total fraction lost) = AEQRecruits
So Age1 = AEQRecruits/(1-total fraction lost)
Set Cohort[1], LastRanError, LastRanFlow, LastRanMarine
SAVE YEAR DATA
stores the year data that will be needed for the statistics at the end
}
\description{
RunSims takes the input list and runs the VRAP functions
}
