% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveBYrData.R
\name{SaveBYrData}
\alias{SaveBYrData}
\title{SaveBYrData}
\usage{
SaveBYrData(inputs, SummaryStats)
}
\arguments{
\item{inputs}{Inputs from .rav file}

\item{SummaryStats}{list with the summary statistics to be updated}
}
\value{
Nothing.  Writes file.
}
\description{
Write the .byr output file
}
