% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveEscpmntData.R
\name{SaveEscpmntData}
\alias{SaveEscpmntData}
\title{SaveEscpmntData}
\usage{
SaveEscpmntData(inputs, SummaryStats)
}
\arguments{
\item{inputs}{Inputs from .rav file}

\item{SummaryStats}{list with the summary statistics to be updated}
}
\value{
Nothing.  Writes file.
}
\description{
Write the .esc output file
}
