% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveSummary.R
\name{SaveSummary}
\alias{SaveSummary}
\title{SaveSummary}
\usage{
SaveSummary(inputs, SummaryStats, staticvars)
}
\arguments{
\item{inputs}{Inputs from .rav file}

\item{SummaryStats}{list with the summary statistics to be updated}

\item{staticvars}{Static variables}
}
\value{
Nothing.  Writes file.
PRINT HEADER INFORMATION
njs MxR adjusted for average environmental conditions
}
\description{
Write the .esc output file
}
