% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveYearData.R
\name{SaveYearData}
\alias{SaveYearData}
\title{SaveYearData}
\usage{
SaveYearData(Year, YearStats)
}
\arguments{
\item{Year}{year}

\item{YearStats}{list of computed variables for each year: AEQMort, Escpmnt[Year,] = Escpmnt, TotAdultEscpmnt, TotAEQMort, TotEscpmnt,TempCohort.}
}
\value{
Updated YearStats list
}
\description{
Update CalendarHR[Year] in YearStats
}
