% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOutFileNames.R
\name{SetOutFileNames}
\alias{SetOutFileNames}
\title{SetOutFileNames}
\usage{
SetOutFileNames(BaseName, inputs, PathName = NULL)
}
\arguments{
\item{BaseName}{The basename for the .sum, .byr, and .esc output files}

\item{inputs}{Inputs from .rav file}

\item{PathName}{Path for the files}
}
\value{
Updated inputs list with full names for the output files
}
\description{
Set output file names
}
\details{
Used by GetOutFiles and also by GetCommandLine
}
