% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupSummaryStats.R
\name{SetupSummaryStats}
\alias{SetupSummaryStats}
\title{SetupSummaryStats}
\usage{
SetupSummaryStats(inputs)
}
\arguments{
\item{inputs}{Inputs from .rav file}
}
\value{
SummaryStats list
}
\description{
Create an list set up for all the summary stats. Each list itme is length of BufMax and is all 0.
}
