% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trend.R
\name{Trend}
\alias{Trend}
\title{Trend}
\usage{
Trend(t, y, x, z)
}
\arguments{
\item{t}{trend rate}

\item{y}{time increment}

\item{x}{first value}

\item{z}{type of trend 0 or 1}
}
\value{
trend variable (scalar or vector)
EEH: changed to work with vectors of y
}
\description{
Compute variable x with trend t for year y
}
\details{
NJS: created 7/9/02. Original VB code:
Function Trend(t As Double, y, x As Double, z) As Double
t is trend rate, y is time increment, x is first value, z is type of trend

If z = 0 Then
Trend = x * (1 + t) ^ y
ElseIf z = 1 Then
Trend = x + (y * t)
Else
If Trend < 0 Then Trend = 0

Print "Unknown trend/cycle function"
Stop
End If

End Function
}
