% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteRavFile.R
\name{WriteRavFile}
\alias{WriteRavFile}
\title{WriteRavFile}
\usage{
WriteRavFile(inputs, filename)
}
\arguments{
\item{inputs}{Inputs from .rav file.}

\item{filename}{Name of the .rav file to write.}
}
\value{
Nothing. .rav file is written.
}
\description{
This function takes inputs list and creates a RAV file
}
\details{
This function is new to the R VRAP.  Needed to create record of the 
.rav file used for the VRAP run in case the user changed .rav values in the
shiny app.
}
