% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReport.R
\name{WriteReport}
\alias{WriteReport}
\title{WriteReport}
\usage{
WriteReport(InFile = NULL, OutFileBase = NULL, show.file = FALSE)
}
\arguments{
\item{InFile}{the .rav input file.}

\item{OutFileBase}{basename to give the outputfiles.}

\item{show.file}{Whether to open the pdf after it is produced.}
}
\value{
Nothing. The pdf is made and saved.
}
\description{
Create a pdf with basic information about the VRAP output.
}
\details{
knit2pdf is used to create the pdf using Report-knitr-ER.xRnw or Report-knitr-Pop.xRnw (Sweave files) in inst/doc.
}
