% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompEscpmnt.R
\name{CompEscpmnt}
\alias{CompEscpmnt}
\title{CompEscpmnt}
\usage{
CompEscpmnt(Regime, Year, inputs, BufTargetU, Cohort, AEQ, YearStats)
}
\arguments{
\item{Regime}{Harvest regime.}

\item{Year}{Year to compute the escapement for.}

\item{inputs}{Inputs from .rav file}

\item{BufTargetU}{Target ER to use for simulation.  Adjusted if StepFunc=ER.}

\item{Cohort}{Cohort}

\item{AEQ}{Adult equivalents}

\item{YearStats}{list of computed variables for each year: AEQMort, Escpmnt[Year,] = Escpmnt, TotAdultEscpmnt, TotAEQMort, TotEscpmnt,TempCohort.}
}
\value{
Updated YearStats list for value of variables in Year
INITIALIZE PARAMETERS, VARIABLES, AND ARRAYS
SET EXPLOITATION RATE TARGET AFTER ADJUSTMENT FOR MANAGEMENT ERROR.
SET INITIAL FISHERY SCALE To 1.0  UNLESS TARGET U=0
COMPUTE PRETERMINAL MORTALITY AND UPDATE COHORT
COMPUTE MATURE RUN AND UPDATE COHORT
COMPUTE MATURE MORTALITY AND ESCAPEMENT
MatU, MatMort, Escpmnt, MatRun are all 1:MaxAge vectors
COMPUTE AEQ TOTAL MORTALITY EXPLOITATION RATE
COMPARE ACTUAL AND TARGET; SCALE EXPLOITATION RATES IF NECESSARY
if any of these met, exit
Criteria 1: unclear why needed.  Criteria 2: target ER < 0. Criteria 3: extinct
}
\description{
Compute escapement
}
