% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompStatsEEH.R
\name{CompStatsEEH}
\alias{CompStatsEEH}
\title{CompStats}
\usage{
CompStatsEEH(BufNum, inputs, BufSRb, YearStats, SummaryStats)
}
\arguments{
\item{BufNum}{Which simulation (TargetER or Pop) is this for.}

\item{inputs}{Inputs from .rav file}

\item{BufSRb}{Capacity.  Changes if StepFunc=Pop.  Otherwise stays the same.}

\item{YearStats}{list of computed variables for each year: AEQMort, Escpmnt[Year,] = Escpmnt, TotAdultEscpmnt, TotAEQMort, TotEscpmnt,TempCohort.}

\item{SummaryStats}{list with the summary statistics to be updated}
}
\value{
Updated SummaryStats list
COMPUTE STATISTICS
Sum up over all reps and divide by reps
COMPUTE AVERAGE MORTALITY OVER REPETITIONS
COMPUTE AVERAGE CALENDAR YEAR HARVEST RATE
COMPUTE PROPORTION OF TIMES ESCAPEMENT WAS LESS THAN LOWER LEVEL
add 1/NRuns because we want the fraction out of all runs
FIND MINIMUM, MAXIMUM, AND AVERAGE ESCAPEMENT        
COMPUTE PROPORTION OF TIMES ENDING ESCAPEMENT WENT BELOW QET (DL2) (depensation)
divide by NRuns at end to get frac of reps below
add 1/NRuns because we want the fraction out of all runs
COMPUTE PROPORTION OF TIMES ENDING ESCAPEMENT EXCEEDED UPPER LEVEL
add 1/NRuns because we want the fraction out of all runs
COMPUTE BROOD YEAR HARVEST RATES
AGE 2
AGE 3
AGE 4
AGE 5
}
\description{
Returns the statistics (calculated values) needed to produce the summary output file
}
\details{
This function similar but not identical to the original VB function
}
