% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInput.R
\name{GetInput}
\alias{GetInput}
\title{GetInput}
\usage{
GetInput(InFile)
}
\arguments{
\item{InFile}{the name of the .rav file}
}
\value{
Returns the list of all inputs
}
\description{
Read in a .rav file and assign all the variables
}
